Refactoring: moved package-level checks into its own module
